requirements.txt
flask==3.0.3
stripe==10.11.0
waitress==3.0.0
python-dotenv==1.0.1